Clean up small bug

+++ app/src/main/java/com/example/toptentrivia/data/model/User.kt
@@ -19,13 +19,13 @@ data class User(
     val username: String,
     val password: String,
 
-    var pointsToday: Double = 0.0, // update after question answered
+    var pointsToday: Int = 0, // update after question answered
     var correctAnswersToday: Int = 0, // update after question answered
     var questionsAttemptedToday: Int = 0, // update at start of question
     var streak: Int = 0, // update on login on new day
     var gamesPlayedAllTime: Int = 0, // update on game start
     var averagePoints: Double = 0.0, // update on login on new day
-    var totalPoints: Double = 0.0, // total points
+    var totalPoints: Int = 0, // total points
     var correctAnswersAllTime: Int = 0, // update after question answered
     var questionsAttemptedAllTime: Int = 0, // update at start of question
     var topTenFinishes: Int = 0, // unused for now
